moving gallery and blog pages out of home dir
