+++ pages/blogs/index.tsx
@@ -1,7 +1,7 @@
 import HeadComponent from '@/components/common/HeadComponent';
 import NavigationBar from '@/components/common/NavigationBar/NavigationBar';
 import BlogsList from '@/components/blogs/BlogsList';
-import blogs from '../../../public/blogs/blogs.json';
+import blogs from '../../public/blogs/blogs.json';
 import type { NextPage, GetStaticProps, InferGetStaticPropsType } from 'next';
 import type { TBlog } from '@/utils/types';
 
@@ -17,7 +17,7 @@ const Blogs: NextPage<InferGetStaticPropsType<typeof getStaticProps>> = ({
   return (
     <>
       <HeadComponent pageTitle='Blogs'></HeadComponent>
-      <div className='min-w-screen flex min-h-screen items-start justify-center bg-primary-theme-bg pt-16'>
+      <div className='min-w-screen flex min-h-screen items-start justify-center pt-16'>
         <NavigationBar defaultValue={1} />
         <BlogsList blogs={blogs}></BlogsList>
       </div>
